refactor(control): colocate nfqueue reconstruction
